Refactored explore resources view page to work separately.
